# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: WavesOfEgypt travel marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "409":
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /destinations:
    get:
      operationId: listDestinations
      tags: [destinations]
      summary: List all destinations
      parameters:
        - name: featured
          in: query
          schema:
            type: boolean
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of destinations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Destination"

    post:
      operationId: createDestination
      tags: [destinations]
      summary: Create a destination (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DestinationInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Destination"
  /destinations/{id}:
    get:
      operationId: getDestination
      tags: [destinations]
      summary: Get destination by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Destination
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Destination"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  # ── CATEGORIES ───────────────────────────────────────────────────────────
...
  /reviews:
    get:
      operationId: listReviews
      tags: [reviews]
      summary: List reviews for a tour
      parameters:
        - name: tourId
          in: query
          required: true
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"

    post:
      operationId: createReview
      tags: [reviews]
      summary: Create a review
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"

  # ── BOOKINGS ─────────────────────────────────────────────────────────────
  /bookings:
    get:
      operationId: listBookings
      tags: [bookings]
      summary: List user bookings (or all if admin/vendor)
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, confirmed, cancelled, completed]
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"

    post:
      operationId: createBooking
      tags: [bookings]
      summary: Create a booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingInput"
      responses:
        "201":
          description: Booking created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /bookings/{id}:
    get:
      operationId: getBooking
      tags: [bookings]
      summary: Get a booking by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"

    patch:
      operationId: updateBooking
      tags: [bookings]
      summary: Update booking status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingUpdate"
      responses:
        "200":
          description: Updated booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"

  # ── WISHLIST ──────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    RegisterInput:
      type: object
      required: [email, password, name]
      properties:
        email:
          type: string
        password:
          type: string
          minLength: 6
        name:
          type: string
        role:
          type: string
          enum: [customer, vendor]
    AuthResponse:
      type: object
      required: [user]
      properties:
        user:
          $ref: "#/components/schemas/User"
    User:
      type: object
      required: [id, email, name, role]
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [customer, vendor, admin]
        avatar:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time

    # ── DESTINATION ─────────────────────────────────────────────────────────
    Destination:
      type: object
      required: [id, name, slug, region, imageUrl, tourCount]
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: ["string", "null"]
        region:
          type: string
          enum: [red_sea, nile_valley, mediterranean, western_desert, sinai]
        imageUrl:
          type: string
        tourCount:
          type: integer
        featured:
          type: boolean
    DestinationInput:
      type: object
      required: [name, slug, region, imageUrl]
      properties:
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        region:
          type: string
          enum: [red_sea, nile_valley, mediterranean, western_desert, sinai]
        imageUrl:
          type: string
        featured:
          type: boolean

    # ── CATEGORY ─────────────────────────────────────────────────────────────
...
    Review:
      type: object
      required: [id, tourId, userId, rating, createdAt]
      properties:
        id:
          type: integer
        tourId:
          type: integer
        userId:
          type: integer
        userName:
          type: ["string", "null"]
        userAvatar:
          type: ["string", "null"]
        rating:
          type: number
        comment:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    ReviewInput:
      type: object
      required: [tourId, rating]
      properties:
        tourId:
          type: integer
        rating:
          type: number
          minimum: 1
          maximum: 5
        comment:
          type: string

    # ── BOOKING ──────────────────────────────────────────────────────────────
    Booking:
      type: object
      required: [id, tourId, userId, date, participants, totalPrice, status, createdAt]
      properties:
        id:
          type: integer
        tourId:
          type: integer
        tourTitle:
          type: ["string", "null"]
        tourCoverImage:
          type: ["string", "null"]
        userId:
          type: integer
        userName:
          type: ["string", "null"]
        date:
          type: string
          format: date
        participants:
          type: integer
        totalPrice:
          type: number
        status:
          type: string
          enum: [pending, confirmed, cancelled, completed]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    BookingInput:
      type: object
      required: [tourId, date, participants]
      properties:
        tourId:
          type: integer
        date:
          type: string
          format: date
        participants:
          type: integer
          minimum: 1
        notes:
          type: string
    BookingUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [pending, confirmed, cancelled, completed]
        notes:
          type: string

    # ── WISHLIST ─────────────────────────────────────────────────────────────
...