# lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * WavesOfEgypt travel marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { BookingStatus } from './bookingStatus';

export interface Booking {
  id: number;
  tourId: number;
  /** @nullable */
  tourTitle?: string | null;
  /** @nullable */
  tourCoverImage?: string | null;
  userId: number;
  /** @nullable */
  userName?: string | null;
  date: Date;
  participants: number;
  totalPrice: number;
  status: BookingStatus;
  /** @nullable */
  notes?: string | null;
  createdAt: Date;
}